backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
supabase==2.9.0
openai==1.52.0
pinecone-client==5.0.1
pdfplumber==0.11.4
redis==5.1.1
celery==5.4.0
sentry-sdk[fastapi]==2.16.0
python-dotenv==1.0.1
